+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.5] - 2026-05-29
 
 ### Added
 - parsed env variables optionally have double underscores replaced by dot
 - explicit side effect in funcflow
+
+### Fixed
+- JsonObject for python 3.14
 
 
 ## [1.0.4] - 2026-04-23